+++ website/docs/releases/release-notes.md
@@ -2,7 +2,7 @@
 title: Release Notes
 ---
 
-## v0.13.0-next.5
+## v0.13.0-next.6
 
 This release contains several breaking changes, please refer to the [migration notes](/docs/releases/migration-notes) for more information.
 
@@ -22,11 +22,14 @@ This release contains several breaking changes, please refer to the [migration n
 
 - :sparkles: Add support for loops in debug tooling.
 - :sparkles: Add support for `p2s` contract type.
+- :sparkles: Add `addBchChangeOutputIfNeeded()` method to `TransactionBuilder` class.
 - :sparkles: Add `lockingBytecode` property to `Contract` class.
 - :sparkles: Add `getUtxosForLockingBytecode()` method to `ElectrumNetworkProvider` class and `MockNetworkProvider` interface.
 - :sparkles: In the `MockNetworkProvider`, `addUtxo()` now also allows UTXOs to be added by locking bytecode.
+- :sparkles: Add `gatherBchUtxos()` and `gatherFungibleTokenUtxos()` functions to the SDK for gathering UTXOs.
 - :hammer_and_wrench: **BREAKING**: Rename `addressType` option on `Contract` constructor to `contractType`.
 - :hammer_and_wrench: **BREAKING**: Remove undocumented `redeemScript` property from `Contract` class.
+- :hammer_and_wrench: **BREAKING**: Remove undocumented `buildLibauthTransaction()` method from `TransactionBuilder` class.
 - :hammer_and_wrench: Update default VM target to `BCH_2026_05`.
 - :hammer_and_wrench: Improve package size by tidying up dependencies.